pyproject.toml
[project]
name = "businessmanagement"
version = "0.1.0"
description = ""
authors = [
    {name = "AleksandrZaec",email = "zaecam95@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.2,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-extensions (>=4.1,<5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
